consolidate environment detection helpers
